feat: add method to url
